+++ CHANGELOG.md
@@ -21,6 +21,7 @@ contents to a new release section (e.g. _[0.x]_).
 - Solve nested logarithmic inequalities. (PLUT-1142)
 - Solve exponential inequalities with one exponential expression. (PLUT-1151)
 - Solve exponential inequalities with two exponential expressions. (PLUT-1154)
+- Coefficient removal for exponential equalities and inequalities. (PLUT-1157)
 - SDK: Add support for exponent on log function. (PLUT-1127)
 
 ### Fixed